Add "merge_day_clobber_file" to Treescript worker config.

+++ treescript/src/treescript/merges.py
@@ -25,20 +25,23 @@ def replace(file_name, from_, to_):
         f.write(new_text)
 
 
-def touch_clobber_file(repo_path):
+def touch_clobber_file(config, repo_path):
     """Update the clobber file in the root of the repo."""
-    log.info("Touching clobber file")
-    clobber_file = os.path.join(repo_path, "CLOBBER")
-    with open(clobber_file) as f:
-        contents = f.read()
-    new_contents = ""
-    for line in contents.splitlines():
-        line = line.strip()
-        if line.startswith("#") or line == "":
-            new_contents += f"{line}\n"
-    new_contents += "Merge day clobber"
-    with open(clobber_file, "w") as f:
-        f.write(new_contents)
+    if config["merge_day_clobber_file"]:
+        log.info("Touching clobber file")
+        clobber_file = os.path.join(repo_path, config["merge_day_clobber_file"])
+        with open(clobber_file) as f:
+            contents = f.read()
+        new_contents = ""
+        for line in contents.splitlines():
+            line = line.strip()
+            if line.startswith("#") or line == "":
+                new_contents += f"{line}\n"
+        new_contents += "Merge day clobber"
+        with open(clobber_file, "w") as f:
+            f.write(new_contents)
+    else:
+        log.info("merge_day_clobber_file not set, skipping clobber file update")
 
 
 async def apply_rebranding(config, repo_path, merge_config):
@@ -62,7 +65,7 @@ async def apply_rebranding(config, repo_path, merge_config):
     for f, from_, to in merge_config.get("replacements", list()):
         replace(os.path.join(repo_path, f), from_, to)
 
-    touch_clobber_file(repo_path)
+    touch_clobber_file(config, repo_path)
 
 
 # do_merge {{{1
